v1.1

+++ src/main/java/dev/drekamor/civminer/config/ConfigInitializer.java
@@ -4,6 +4,7 @@ import com.google.gson.JsonObject;
 import dev.drekamor.civminer.CivMiner;
 import net.fabricmc.loader.api.FabricLoader;
 
+import java.io.File;
 import java.io.FileWriter;
 import java.io.IOException;
 import java.nio.file.Path;
@@ -12,20 +13,24 @@ public class ConfigInitializer {
     private static Path getPath = FabricLoader.getInstance().getConfigDir();
     private static String path = getPath + "/civminer.json";
 
-    public static void init(){
+    public static void init() {
         String stop_text = "You sense debris";
-        int stop_distance = 1000;
+        String message = "/g Icarus <@284408888109957130>";
         Boolean safe_mode = true;
         JsonObject config = new JsonObject();
         config.addProperty("stop_text", stop_text);
-        config.addProperty("stop_distance", stop_distance);
+        config.addProperty("message", message);
         config.addProperty("safe_mode", safe_mode);
         String text = CivMiner.gson.toJson(config);
-        try {
-            FileWriter writer = new FileWriter(path);
-            writer.write(text);
-            writer.close();
-        } catch (IOException e) {
+        File file = new File(path);
+        if (!file.exists()) {
+            try {
+                FileWriter writer = new FileWriter(path);
+                writer.write(text);
+                writer.close();
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
         }
     }
 }
